requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.15
langchain-neo4j>=0.1.1
openai>=1.2.0
neo4j>=5.14.0
python-dotenv>=1.0.0
uuid>=1.30
zep-cloud>=0.1.0
asyncio>=3.4.3
pandas>=2.0.0
